# openapi.yaml
openapi: 3.0.3
info:
  title: CryptoTradeManager Authentication API
  description: |
    CryptoTradeManager 암호화폐 거래 플랫폼을 위한 인증 시스템 API
    
    **주요 기능:**
    - JWT 토큰 기반 인증 (Access + Refresh Token)
    - OAuth2.0 소셜 로그인 (Google, Apple)
    - 2단계 인증 (2FA) 지원
    - Role 기반 접근 제어 (RBAC)
    - 거래소 API 키 관리
    - WebSocket 인증
    - Rate Limiting
  version: 1.0.0
  contact:
    name: CryptoTradeManager Team
    email: api@cryptotrademanager.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cryptotrademanager.com/v1
    description: Production server
  - url: https://staging-api.cryptotrademanager.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Access Token 갱신
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 토큰 갱신 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
...
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 3600
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: 요청이 유효하지 않습니다.
        details:
          type: object
          additionalProperties:
            type: string
        timestamp:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00Z
        path:
          type: string
          example: /v1/auth/signin
...
  responses:
...
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: 인증이 필요합니다.
            timestamp: 2024-01-01T12:00:00Z
            path: /v1/auth/user/profile
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 토큰을 사용한 인증. 
        Authorization 헤더에 "Bearer {token}" 형식으로 전송합니다.
        
        **토큰 구조:**
        - Header: {"alg": "HS256", "typ": "JWT"}
        - Payload: {"sub": "user_id", "iat": timestamp, "exp": timestamp, "roles": ["USER"]}
        - Signature: HMACSHA256(base64UrlEncode(header) + "." + base64UrlEncode(payload), secret)
...